+++ dotcom-rendering/src/components/ShowMore.importable.tsx
@@ -14,11 +14,12 @@ import {
 	SvgPlus,
 } from '@guardian/source/react-components';
 import { useEffect, useState } from 'react';
+import type { FEFrontCard } from '../frontend/feFront';
 import { shouldPadWrappableRows } from '../lib/dynamicSlices';
 import type { EditionId } from '../lib/edition';
 import { useApi } from '../lib/useApi';
 import { enhanceCards } from '../model/enhanceCards';
-import type { DCRContainerPalette, FEFrontCard } from '../types/front';
+import type { DCRContainerPalette } from '../types/front';
 import { LI } from './Card/components/LI';
 import { UL } from './Card/components/UL';
 import { FrontCard } from './FrontCard';
