+++ dotcom-rendering/src/frontend/feTagPage.ts
@@ -1,9 +1,9 @@
 import type { EditionId } from '../lib/edition';
 import type { CommercialProperties } from '../types/commercial';
 import type { FooterType } from '../types/footer';
-import type { FEFrontCard, FEFrontConfigType } from '../types/front';
 import type { FENavType } from '../types/frontend';
 import type { FEPagination, FETagType } from '../types/tag';
+import type { FEFrontCard, FEFrontConfig } from './feFront';
 
 export type FETagPage = {
 	contents: FEFrontCard[];
@@ -18,7 +18,7 @@ export type FETagPage = {
 	pageId: string;
 	webTitle: string;
 	webURL: string;
-	config: FEFrontConfigType;
+	config: FEFrontConfig;
 	commercialProperties: CommercialProperties;
 	pageFooter: FooterType;
 	isAdFreeUser: boolean;
