+++ dotcom-rendering/src/lib/contributions.ts
@@ -9,7 +9,7 @@ import { useEffect, useState } from 'react';
 import { hideSupportMessaging } from '../client/userFeatures/cookies/hideSupportMessaging';
 import { userBenefitsDataIsUpToDate } from '../client/userFeatures/cookies/userBenefitsExpiry';
 import type { ArticleDeprecated } from '../types/article';
-import type { DCRFrontType } from '../types/front';
+import type { Front } from '../types/front';
 import type { DCRNewslettersPageType } from '../types/newslettersPage';
 import type { TagPage } from '../types/tagPage';
 
@@ -195,7 +195,7 @@ export const recentlyClosedBanner = (
 };
 
 export const getContributionsServiceUrl = (
-	config: ArticleDeprecated | DCRFrontType | TagPage | DCRNewslettersPageType,
+	config: ArticleDeprecated | Front | TagPage | DCRNewslettersPageType,
 ): string => process.env.SDC_URL ?? config.contributionsServiceUrl;
 
 type PurchaseInfo = HeaderPayload['targeting']['purchaseInfo'];
