+++ dotcom-rendering/src/model/enhanceCards.ts
@@ -1,4 +1,9 @@
 import { isUndefined } from '@guardian/libs';
+import type {
+	FEFrontCard,
+	FEMediaAtom,
+	FESupportingContent,
+} from '../frontend/feFront';
 import {
 	ArticleDesign,
 	type ArticleFormat,
@@ -12,9 +17,6 @@ import type {
 	DCRFrontCard,
 	DCRSlideshowImage,
 	DCRSupportingContent,
-	FEFrontCard,
-	FEMediaAtom,
-	FESupportingContent,
 } from '../types/front';
 import type { MainMedia } from '../types/mainMedia';
 import type { PodcastSeriesImage, TagType } from '../types/tag';
