+++ dotcom-rendering/src/model/enhanceCollections.ts
@@ -1,7 +1,8 @@
+import type { FECollection } from '../frontend/feFront';
 import { decideCollectionBranding } from '../lib/branding';
 import type { EditionId } from '../lib/edition';
 import type { Branding } from '../types/branding';
-import type { DCRCollectionType, FECollectionType } from '../types/front';
+import type { DCRCollectionType } from '../types/front';
 import { decideContainerPalette } from './decideContainerPalette';
 import { enhanceCards } from './enhanceCards';
 import { enhanceTreats } from './enhanceTreats';
@@ -16,7 +17,7 @@ const FORBIDDEN_CONTAINERS = [
 const PALETTE_STYLES_URI =
 	'https://content.guardianapis.com/atom/interactive/interactives/2022/03/29/fronts-container-colours/default';
 
-const isSupported = (collection: FECollectionType): boolean =>
+const isSupported = (collection: FECollection): boolean =>
 	!(
 		FORBIDDEN_CONTAINERS.includes(collection.displayName) ||
 		collection.curated.some(
@@ -25,7 +26,7 @@ const isSupported = (collection: FECollectionType): boolean =>
 	);
 
 const findCollectionSuitableForFrontBranding = (
-	collections: FECollectionType[],
+	collections: FECollection[],
 ) => {
 	// Find the lowest indexed collection that COULD display branding
 	const index = collections.findIndex(
@@ -43,7 +44,7 @@ const findCollectionSuitableForFrontBranding = (
 };
 
 /** Depending on the next sibling of the container, we assign either large or small spacing rules during render */
-const getContainerSpacing = (nextSiblingCollection?: FECollectionType) => {
+const getContainerSpacing = (nextSiblingCollection?: FECollection) => {
 	const nextCollectionIsPrimary =
 		nextSiblingCollection?.config.collectionLevel === 'Primary';
 	return nextCollectionIsPrimary ? 'large' : 'small';
@@ -58,7 +59,7 @@ export const enhanceCollections = ({
 	onPageDescription,
 	isOnPaidContentFront,
 }: {
-	collections: FECollectionType[];
+	collections: FECollection[];
 	editionId: EditionId;
 	pageId: string;
 	discussionApiUrl: string;
