+++ dotcom-rendering/src/types/front.ts
@@ -1,46 +1,32 @@
-import type { SharedAdTargeting } from '../lib/ad-targeting';
+import type {
+	FEAspectRatio,
+	FEContainer,
+	FEContainerLevel,
+	FEFrontConfig,
+	FEFrontProperties,
+	FESeoData,
+} from '../frontend/feFront';
 import type {
 	ArticleFormat,
 	ArticleSpecial,
 	Pillar,
 } from '../lib/articleFormat';
 import type { EditionId } from '../lib/edition';
-import type { Branding, CollectionBranding, EditionBranding } from './branding';
-import type { ServerSideTests, StageType, Switches } from './config';
-import type { BoostLevel, Image, StarRating } from './content';
+import type { Branding, CollectionBranding } from './branding';
+import type { BoostLevel, StarRating } from './content';
 import type { FooterType } from './footer';
-import type { FEFormat, FENavType } from './frontend';
+import type { FENavType } from './frontend';
 import type { MainMedia } from './mainMedia';
 import type { FETagType } from './tag';
 import type { Territory } from './territory';
-import type { FETrailType, TrailType } from './trails';
+import type { TrailType } from './trails';
 
-export interface FEFrontType {
-	pressedPage: FEPressedPageType;
-	nav: FENavType;
-	editionId: EditionId;
-	editionLongForm: string;
-	guardianBaseURL: string;
-	pageId: string;
-	webTitle: string;
-	webURL: string;
-	config: FEFrontConfigType;
-	commercialProperties: Record<string, unknown>;
-	pageFooter: FooterType;
-	isAdFreeUser: boolean;
-	isNetworkFront: boolean;
-	mostViewed: FETrailType[];
-	deeplyRead?: FETrailType[];
-	contributionsServiceUrl: string;
-	canonicalUrl?: string;
-}
-
-export interface DCRFrontType {
-	pressedPage: DCRPressedPageType;
+export interface Front {
+	pressedPage: PressedPage;
 	nav: FENavType;
 	editionId: EditionId;
 	webTitle: string;
-	config: FEFrontConfigType;
+	config: FEFrontConfig;
 	pageFooter: FooterType;
 	isAdFreeUser: boolean;
 	isNetworkFront: boolean;
@@ -54,90 +40,12 @@ export interface DCRFrontType {
 	guardianBaseURL: string;
 }
 
-interface FEPressedPageType {
+interface PressedPage {
 	id: string;
-	seoData: FESeoDataType;
-	frontProperties: FEFrontPropertiesType;
-	collections: FECollectionType[];
-}
-
-interface DCRPressedPageType {
-	id: string;
-	seoData: FESeoDataType;
-	frontProperties: FEFrontPropertiesType;
+	seoData: FESeoData;
+	frontProperties: FEFrontProperties;
 	collections: DCRCollectionType[];
 }
-
-/* This list of containers supported in DCR must be kept up to date with frontend **manually**.
- * @see https://github.com/guardian/frontend/blob/167dce23a8453ed13a97fbd23c7fc45ecb06e3fe/facia/app/services/dotcomrendering/FaciaPicker.scala#L21-L45 */
-type FEContainerType =
-	| 'dynamic/fast'
-	| 'dynamic/package'
-	| 'dynamic/slow'
-	| 'dynamic/slow-mpu'
-	| 'fixed/large/slow-XIV'
-	| 'fixed/medium/fast-XI'
-	| 'fixed/medium/fast-XII'
-	| 'fixed/medium/slow-VI'
-	| 'fixed/medium/slow-VII'
-	| 'fixed/medium/slow-XII-mpu'
-	| 'fixed/small/fast-VIII'
-	| 'fixed/small/slow-I'
-	| 'fixed/small/slow-III'
-	| 'fixed/small/slow-IV'
-	| 'fixed/small/slow-V-half'
-	| 'fixed/small/slow-V-mpu'
-	| 'fixed/small/slow-V-third'
-	| 'fixed/thrasher'
-	| 'fixed/video'
-	| 'fixed/video/vertical'
-	| 'nav/list'
-	| 'nav/media-list'
-	| 'news/most-popular'
-	| 'scrollable/highlights'
-	| 'flexible/special'
-	| 'flexible/general'
-	| 'scrollable/small'
-	| 'scrollable/medium'
-	| 'scrollable/feature'
-	| 'static/feature/2'
-	| 'static/medium/4';
-
-export type FEContainerLevel = 'Primary' | 'Secondary';
-
-export type FEContainerMetadata =
-	| 'EventPalette'
-	| 'SombreAltPalette'
-	| 'EventAltPalette'
-	| 'InvestigationPalette'
-	| 'LongRunningAltPalette'
-	| 'LongRunningPalette'
-	| 'SombrePalette'
-	| 'Canonical'
-	| 'Dynamo'
-	| 'Special'
-	| 'DynamoLike'
-	| 'Breaking'
-	| 'Podcast'
-	| 'Branded'
-	| 'BreakingPalette'
-	| 'SpecialReportAltPalette'
-	| 'Secondary';
-
-export type FEFrontCardStyle =
-	| 'SpecialReport'
-	| 'SpecialReportAlt'
-	| 'LiveBlog'
-	| 'DeadBlog'
-	| 'Feature'
-	| 'Editorial'
-	| 'Comment'
-	| 'Media'
-	| 'Analysis'
-	| 'Review'
-	| 'Letters'
-	| 'ExternalLink'
-	| 'DefaultCardstyle';
 
 export type DCRContainerPalette =
 	| 'EventPalette'
@@ -154,166 +62,9 @@ export type DCRContainerPalette =
 	| 'PodcastPalette';
 
 // TODO: These may need to be declared differently than the front types in the future
-export type DCRContainerType = FEContainerType;
+export type DCRContainerType = FEContainer;
 
 export type DCRContainerLevel = FEContainerLevel;
-
-/** @see https://github.com/guardian/frontend/blob/0bf69f55a/common/app/model/content/Atom.scala#L191-L196 */
-interface MediaAsset {
-	id: string;
-	version: number;
-	platform: string;
-	mimeType?: string;
-}
-
-/** @see https://github.com/guardian/frontend/blob/0bf69f55a/common/app/model/content/Atom.scala#L158-L169 */
-export interface FEMediaAtom {
-	id: string;
-	// defaultHtml: string; // currently unused
-	assets: MediaAsset[];
-	title: string;
-	duration?: number;
-	source?: string;
-	posterImage?: { allImages: Image[] };
-	expired?: boolean;
-	activeVersion?: number;
-	// channelId?: string; // currently unused
-}
-
-export type FEFrontCard = {
-	properties: {
-		isBreaking: boolean;
-		showMainVideo: boolean;
-		showKickerTag: boolean;
-		showByline: boolean;
-		imageSlideshowReplace: boolean;
-		maybeContent?: {
-			trail: {
-				trailPicture?: {
-					allImages: {
-						index: number;
-						fields: {
-							displayCredit?: string;
-							source?: string;
-							photographer?: string;
-							isMaster?: string;
-							altText?: string;
-							height: string;
-							credit?: string;
-							mediaId?: string;
-							width: string;
-						};
-						mediaType: string;
-						url: string;
-					}[];
-				};
-				byline?: string;
-				thumbnailPath?: string;
-				webPublicationDate: number;
-			};
-			metadata: {
-				id: string;
-				webTitle: string;
-				webUrl: string;
-				type: string;
-				sectionId?: { value: string };
-				format: FEFormat;
-			};
-			fields: {
-				main: string;
-				body: string;
-				standfirst?: string;
-			};
-			elements: {
-				mainVideo?: unknown;
-				mediaAtoms: FEMediaAtom[];
-				mainMediaAtom?: FEMediaAtom;
-			};
-			tags: { tags: FETagType[] };
-		};
-		maybeContentId?: string;
-		isLiveBlog: boolean;
-		isCrossword: boolean;
-		byline?: string;
-		image?: {
-			type: string;
-			item: {
-				imageSrc?: string;
-				assets?: {
-					imageSrc: string;
-					imageCaption?: string;
-				}[];
-			};
-		};
-		webTitle: string;
-		linkText?: string;
-		webUrl?: string;
-		editionBrandings: EditionBranding[];
-		href?: string;
-		embedUri?: string;
-	};
-	header: {
-		isVideo: boolean;
-		isComment: boolean;
-		isGallery: boolean;
-		isAudio: boolean;
-		kicker?: {
-			type: string;
-			item?: {
-				properties: {
-					kickerText: string;
-				};
-			};
-		};
-		seriesOrBlogKicker?: {
-			properties: {
-				kickerText: string;
-			};
-			name: string;
-			url: string;
-			id: string;
-		};
-		headline: string;
-		url: string;
-		hasMainVideoElement: boolean;
-	};
-	card: {
-		id: string;
-		cardStyle: {
-			type: FEFrontCardStyle;
-		};
-		webPublicationDateOption?: number;
-		lastModifiedOption?: number;
-		trailText?: string;
-		starRating?: StarRating;
-		shortUrlPath?: string;
-		shortUrl: string;
-		group: string;
-		isLive: boolean;
-		galleryCount?: number;
-		audioDuration?: string;
-	};
-	discussion: {
-		isCommentable: boolean;
-		isClosedForComments: boolean;
-		discussionId?: string;
-	};
-	display: {
-		isBoosted: boolean;
-		boostLevel?: BoostLevel;
-		showBoostedHeadline: boolean;
-		showQuotedHeadline: boolean;
-		imageHide: boolean;
-		showLivePlayable: boolean;
-	};
-	format?: FEFormat;
-	enriched?: FESnapType;
-	supportingContent?: FESupportingContent[];
-	cardStyle?: {
-		type: FEFrontCardStyle;
-	};
-	type: string;
-};
 
 export type DCRFrontImage = {
 	src: string;
@@ -358,61 +109,13 @@ export type DCRSlideshowImage = {
 	imageCaption?: string;
 };
 
-export type FESnapType = {
-	embedHtml?: string;
-	embedCss?: string;
-	embedJs?: string;
-};
-
 export type DCRSnapType = {
 	embedHtml?: string;
 	embedCss?: string;
 	embedJs?: string;
 };
 
-export type AspectRatio = '5:3' | '5:4' | '4:5' | '1:1';
-
-type FECollectionConfigType = {
-	displayName: string;
-	metadata?: { type: FEContainerMetadata }[];
-	collectionType: FEContainerType;
-	collectionLevel?: FEContainerLevel;
-	href?: string;
-	groups?: string[];
-	uneditable: boolean;
-	showTags: boolean;
-	showSections: boolean;
-	hideKickers: boolean;
-	showDateHeader: boolean;
-	showLatestUpdate: boolean;
-	excludeFromRss: boolean;
-	showTimestamps: boolean;
-	hideShowMore: boolean;
-	platform: string;
-	aspectRatio?: AspectRatio;
-};
-
-export type FECollectionType = {
-	id: string;
-	displayName: string;
-	description?: string;
-	curated: FEFrontCard[];
-	backfill: FEFrontCard[];
-	treats: FEFrontCard[];
-	lastUpdate?: number;
-	href?: string;
-	groups?: string[];
-	collectionType: FEContainerType;
-	uneditable: boolean;
-	showTags: boolean;
-	showSections: boolean;
-	hideKickers: boolean;
-	showDateHeader: boolean;
-	showLatestUpdate: boolean;
-	config: FECollectionConfigType;
-	hasMore: boolean;
-	targetedTerritory?: Territory;
-};
+export type AspectRatio = FEAspectRatio;
 
 export type DCRCollectionType = {
 	id: string;
@@ -449,124 +152,6 @@ export type DCRGroupedTrails = {
 	big: DCRFrontCard[];
 	standard: DCRFrontCard[];
 	splash: DCRFrontCard[];
-};
-
-export type FEFrontConfigType = {
-	avatarApiUrl: string;
-	externalEmbedHost: string;
-	ajaxUrl: string;
-	keywords: string;
-	revisionNumber: string;
-	isProd: boolean;
-	switches: Switches;
-	section: string;
-	keywordIds: string;
-	locationapiurl: string;
-	sharedAdTargeting: SharedAdTargeting;
-	buildNumber: string;
-	abTests: ServerSideTests;
-	pbIndexSites: { [key: string]: unknown }[];
-	ampIframeUrl: string;
-	beaconUrl: string;
-	host: string;
-	brazeApiKey?: string;
-	calloutsUrl: string;
-	requiresMembershipAccess: boolean;
-	onwardWebSocket: string;
-	a9PublisherId: string;
-	contentType: string;
-	facebookIaAdUnitRoot: string;
-	ophanEmbedJsUrl: string;
-	idUrl: string;
-	dcrSentryDsn: string;
-	isFront: true;
-	idWebAppUrl: string;
-	discussionApiUrl: string;
-	sentryPublicApiKey: string;
-	omnitureAccount: string;
-	dfpAccountId: string;
-	pageId: string;
-	forecastsapiurl: string;
-	assetsPath: string;
-	pillar: string;
-	commercialBundleUrl: string;
-	discussionApiClientHeader: string;
-	membershipUrl: string;
-	dfpHost: string;
-	cardStyle?: string;
-	googletagUrl: string;
-	sentryHost: string;
-	shouldHideAdverts: boolean;
-	mmaUrl: string;
-	membershipAccess: string;
-	isPreview: boolean;
-	googletagJsUrl: string;
-	supportUrl: string;
-	edition: string;
-	ipsosTag: string;
-	ophanJsUrl: string;
-	isPaidContent?: boolean;
-	mobileAppsAdUnitRoot: string;
-	plistaPublicApiKey: string;
-	frontendAssetsFullURL: string;
-	googleSearchId: string;
-	allowUserGeneratedContent: boolean;
-	dfpAdUnitRoot: string;
-	idApiUrl: string;
-	omnitureAmpAccount: string;
-	adUnit: string;
-	hasPageSkin: boolean;
-	webTitle: string;
-	stripePublicToken: string;
-	googleRecaptchaSiteKey: string;
-	discussionD2Uid: string;
-	googleSearchUrl: string;
-	optimizeEpicUrl: string;
-	stage: StageType;
-	idOAuthUrl: string;
-	isSensitive: boolean;
-	isDev: boolean;
-	thirdPartyAppsAccount?: string;
-	avatarImagesUrl: string;
-	fbAppId: string;
-};
-
-type FESeoDataType = {
-	id: string;
-	navSection: string;
-	webTitle: string;
-	title?: string;
-	description: string;
-};
-
-type FEFrontPropertiesType = {
-	isImageDisplayed: boolean;
-	commercial: {
-		editionBrandings: EditionBranding[];
-		editionAdTargetings: unknown;
-		prebidIndexSites?: unknown;
-	};
-	isPaidContent?: boolean;
-	onPageDescription?: string;
-};
-
-export type FESupportingContent = {
-	properties: {
-		href?: string;
-		webUrl?: string;
-	};
-	header: {
-		kicker?: {
-			item?: {
-				properties: {
-					kickerText: string;
-				};
-			};
-		};
-		headline: string;
-		url: string;
-	};
-	format?: FEFormat;
 };
 
 export type DCRSupportingContent = {
